requirements.txt
SpeechRecognition
openai
python-dotenv
pyaudio
pipwin
LocalAudioPlayer
streamlit